TIP-860: move mass edit processors
