+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/Processor/MassEdit/RemoveProductValueProcessor.php
@@ -1,33 +1,37 @@
 <?php
 
-namespace Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit\Product;
+namespace Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\MassEdit;
 
-use Akeneo\Tool\Component\StorageUtils\Updater\PropertyAdderInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelInterface;
+use Akeneo\Tool\Component\StorageUtils\Updater\PropertyRemoverInterface;
 use Pim\Bundle\EnrichBundle\Connector\Processor\AbstractProcessor;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 /**
- * Processor to add product value in a mass edit
+ * Processor to remove product value in a mass edit
  *
- * @author    Olivier Soulet <olivier.soulet@akeneo.com>
- * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
+ * @author    Philippe Mossière <philippe.mossiere@akeneo.com>
+ * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class AddProductValueProcessor extends AbstractProcessor
+class RemoveProductValueProcessor extends AbstractProcessor
 {
-    /** @var PropertyAdderInterface */
-    protected $propertyAdder;
+    /** @var PropertyRemoverInterface */
+    protected $propertyRemover;
 
     /** @var ValidatorInterface */
     protected $validator;
 
     /**
-     * @param PropertyAdderInterface              $propertyAdder
-     * @param ValidatorInterface                  $validator
+     * @param PropertyRemoverInterface $propertyRemover
+     * @param ValidatorInterface       $validator
      */
-    public function __construct(PropertyAdderInterface $propertyAdder, ValidatorInterface $validator)
-    {
-        $this->propertyAdder = $propertyAdder;
+    public function __construct(
+        PropertyRemoverInterface $propertyRemover,
+        ValidatorInterface $validator
+    ) {
+        $this->propertyRemover = $propertyRemover;
         $this->validator = $validator;
     }
 
@@ -37,7 +41,7 @@ class AddProductValueProcessor extends AbstractProcessor
     public function process($product)
     {
         $actions = $this->getConfiguredActions();
-        $this->addData($product, $actions);
+        $this->removeValuesFromProduct($product, $actions);
 
         if (!$this->isProductValid($product)) {
             $this->stepExecution->incrementSummaryInfo('skipped_products');
@@ -64,15 +68,15 @@ class AddProductValueProcessor extends AbstractProcessor
     }
 
     /**
-     * Add data from $actions to the given $product
+     * Set data from $actions to the given $product
      *
      * @param ProductInterface|ProductModelInterface $product
      * @param array                                  $actions
      */
-    protected function addData($product, array $actions)
+    protected function removeValuesFromProduct($product, array $actions)
     {
         foreach ($actions as $action) {
-            $this->propertyAdder->addData($product, $action['field'], $action['value']);
+            $this->propertyRemover->removeData($product, $action['field'], $action['value']);
         }
     }
 }
