+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Connector/Processor/MassEdit/AddToExistingProductModelProcessorSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit\Product;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\MassEdit;
 
 use Akeneo\Pim\Enrichment\Component\Product\Exception\InvalidArgumentException;
 use Akeneo\Tool\Component\Batch\Job\JobParameters;
