+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Connector/Processor/MassEdit/ChangeParentProcessorSpec.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace spec\Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit\Product;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\MassEdit;
 
 use Akeneo\Tool\Component\Batch\Job\JobParameters;
 use Akeneo\Tool\Component\Batch\Model\JobExecution;
