+++ src/Pim/Bundle/EnrichBundle/Resources/config/connector/processors.yml
@@ -1,11 +1,11 @@
 parameters:
-    pim_enrich.connector.processor.mass_edit.product.update_value.class:                       Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit\Product\UpdateProductValueProcessor
-    pim_enrich.connector.processor.mass_edit.product.add_value.class:                          Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit\Product\AddProductValueProcessor
-    pim_enrich.connector.processor.mass_edit.product.remove_value.class:                       Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit\Product\RemoveProductValueProcessor
-    pim_enrich.connector.processor.mass_edit.product.edit_common_attributes.class:             Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit\Product\EditCommonAttributesProcessor
-    pim_enrich.connector.processor.mass_edit.product.add_attribute_value.class:               Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit\Product\AddAttributeValueProcessor
-    pim_enrich.connector.processor.mass_edit.product.add_to_existing_product_model.class:      Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit\Product\AddToExistingProductModelProcessor
-    pim_enrich.connector.processor.mass_edit.product.edit_attributes.class:        Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit\Product\EditAttributesProcessor
+    pim_enrich.connector.processor.mass_edit.product.update_value.class:                       Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\MassEdit\UpdateProductValueProcessor
+    pim_enrich.connector.processor.mass_edit.product.add_value.class:                          Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\MassEdit\AddProductValueProcessor
+    pim_enrich.connector.processor.mass_edit.product.remove_value.class:                       Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\MassEdit\RemoveProductValueProcessor
+    pim_enrich.connector.processor.mass_edit.product.edit_common_attributes.class:             Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\MassEdit\EditCommonAttributesProcessor
+    pim_enrich.connector.processor.mass_edit.product.add_attribute_value.class:               Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\MassEdit\AddAttributeValueProcessor
+    pim_enrich.connector.processor.mass_edit.product.add_to_existing_product_model.class:      Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\MassEdit\AddToExistingProductModelProcessor
+    pim_enrich.connector.processor.mass_edit.product.edit_attributes.class:        Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\MassEdit\EditAttributesProcessor
     pim_enrich.connector.processor.mass_edit.family.set_requirements.class:                    Akeneo\Pim\Structure\Component\Processor\MassEdit\SetAttributeRequirements
     pim_enrich.connector.processor.quick_export.product_and_product_model.class:               Pim\Bundle\EnrichBundle\Connector\Processor\QuickExport\ProductAndProductModelProcessor
 
@@ -96,7 +96,7 @@ services:
             - '@pim_connector.processor.bulk_media_fetcher'
 
     pim_enrich.connector.processor.mass_edit.product.change_parent:
-        class: Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit\Product\ChangeParentProcessor
+        class: Akeneo\Pim\Enrichment\Component\Product\Connector\Processor\MassEdit\ChangeParentProcessor
         arguments:
             - '@pim_catalog.validator.product'
             - '@pim_catalog.validator.product_model'
